requirements.txt
Sphinx>=1.1.3
Werkzeug>=0.6.2
coverage>=3.5.1
django-debug-toolbar==0.9.4
django-dynamic-fixture==1.4.2
django-extensions==0.9
docutils==0.8.1
flake8==0.8
django-nose==1.1
WebTest==1.3.4
django-webtest==1.5.3
pyflakes==0.5.0
tox==1.4.2
ipdb==0.2
ipython==0.13
South==0.7.6
pinocchio==0.3.1
nose-progressive==1.3
Whoosh==2.4.1
django-oscar-testsupport==0.2
